% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_prepare_data.r
\name{nm_add_date_variables}
\alias{nm_add_date_variables}
\title{Add Date Variables}
\usage{
nm_add_date_variables(df)
}
\arguments{
\item{df}{Input data frame.}
}
\value{
Data frame with added date variables.
}
\description{
\code{nm_add_date_variables} adds date-related variables to the data frame.
}
\examples{
\donttest{
library(dplyr)
library(lubridate)
df <- data.frame(date = Sys.time() + seq(1, 1000, by = 100))
date_added_df <- nm_add_date_variables(df)
}

}
