% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_prepare_data.r
\name{nm_check_data}
\alias{nm_check_data}
\title{Check Data Validity}
\usage{
nm_check_data(df, value)
}
\arguments{
\item{df}{Input data frame.}

\item{value}{The target variable name as a string.}
}
\value{
Data frame with renamed target variable.
}
\description{
\code{nm_check_data} checks the validity of the input data frame and ensures the target variable exists.
}
\examples{
\dontrun{
df <- data.frame(
  target_variable = rnorm(100),
  other_variable = rnorm(100),
  date = Sys.time() + seq(1, 1000, by = 10)
)
checked_df <- nm_check_data(df, value = "target_variable")
}

}
