% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_do_all_unc.r
\name{nm_do_all_unc}
\alias{nm_do_all_unc}
\title{Perform All Steps for Meteorological normalisation with Uncertainty Estimation}
\usage{
nm_do_all_unc(
  df = NULL,
  value = "value",
  feature_names = NULL,
  variables_resample = NULL,
  split_method = "random",
  fraction = 0.75,
  model_config = NULL,
  n_samples = 300,
  n_models = 10,
  confidence_level = 0.95,
  seed = 7654321,
  n_cores = NULL,
  weather_df = NULL,
  memory_save = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{df}{Data frame containing the input data.}

\item{value}{The target variable name as a string.}

\item{feature_names}{The names of the features used for training and normalisation.}

\item{variables_resample}{The names of the variables to be resampled for normalisation. Default is NULL (all feature names except date_unix).}

\item{split_method}{The method for splitting data into training and testing sets. Default is 'random'.}

\item{fraction}{The proportion of the data to be used for training. Default is 0.75.}

\item{model_config}{A list containing configuration parameters for model training.}

\item{n_samples}{Number of samples to generate for normalisation. Default is 300.}

\item{n_models}{Number of models to train for uncertainty estimation. Default is 10.}

\item{confidence_level}{The confidence level for uncertainty estimation. Default is 0.95.}

\item{seed}{A random seed for reproducibility. Default is 7654321.}

\item{n_cores}{Number of CPU cores to use for parallel processing. Default is system's total minus one.}

\item{weather_df}{Optional data frame containing weather data for resampling.}

\item{memory_save}{Logical indicating whether to save memory by processing each sample independently.
If \code{TRUE}, resampling and prediction are done in memory-efficient batches. If \code{FALSE}, all samples
are generated and processed at once, which uses more memory. Default is FALSE.}

\item{verbose}{Should the function print progress messages? Default is TRUE.}
}
\value{
A list containing the normalised data frame with uncertainty estimation and model statistics.
}
\description{
\code{nm_do_all_unc} performs the entire process of training multiple models, normalising the data, and collecting model statistics with uncertainty estimation.
}
\examples{
\dontrun{
library(dplyr)
library(lubridate)
library(progress)
library(purrr)
df <- data.frame(date = Sys.time() + seq(1, 100, by = 1),
                 pollutant = rnorm(100), temp = rnorm(100), humidity = rnorm(100))
result <- nm_do_all_unc(df, value = "pollutant", feature_names = c("temp", "humidity"), n_samples = 300, n_models = 10, seed = 12345)
}
}
