% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_extract_feature_names.r
\name{nm_extract_feature_names}
\alias{nm_extract_feature_names}
\title{Extract and Sort Feature Names by Importance}
\usage{
nm_extract_feature_names(model, importance_ascending = FALSE)
}
\arguments{
\item{model}{The trained H2O model object.}

\item{importance_ascending}{A logical value indicating whether to sort feature names in ascending order
of importance. Default is \code{FALSE} (descending order).}
}
\value{
A vector of sorted feature names based on their importance.
}
\description{
\code{extract_feature_names} extracts feature names from an H2O model and sorts them by their importance.
}
\details{
This function uses H2O's \code{varimp} function to extract feature importance from the model and
returns the feature names sorted by their relative importance. You can control whether the
sorting is in ascending or descending order using the \code{importance_ascending} argument.
}
\examples{
\dontrun{
library(h2o)
h2o.init()
df <- as.h2o(iris)
model <- h2o.gbm(x = 1:4, y = 5, training_frame = df)
feature_names <- extract_feature_names(model, importance_ascending = TRUE)
print(feature_names)
}
}
