% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_normalise.r
\name{nm_generate_resampled}
\alias{nm_generate_resampled}
\title{Generate Resampled Data}
\usage{
nm_generate_resampled(
  df_batch,
  variables_resample,
  replace,
  seed,
  weather_df = NULL
)
}
\arguments{
\item{df_batch}{Input data frame batch.}

\item{variables_resample}{A vector of variables to be resampled.}

\item{replace}{Logical indicating whether to sample with replacement.}

\item{seed}{A random seed for reproducibility.}

\item{weather_df}{Optional data frame containing weather data for resampling.}
}
\value{
A data frame with resampled variables.
}
\description{
\code{nm_generate_resampled} resamples specified variables in the input DataFrame.
}
\examples{
\dontrun{
df <- data.frame(temp = rnorm(100), humidity = rnorm(100))
weather_data <- data.frame(temp = rnorm(100), humidity = rnorm(100))
resampled_df <- nm_generate_resampled(df, variables_resample = c("temp", "humidity"), replace = TRUE, seed = 12345, weather_df = weather_data)
}

}
