% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_save_h2o.r
\name{nm_load_h2o}
\alias{nm_load_h2o}
\title{Load Saved H2O Model}
\usage{
nm_load_h2o(path = "./", filename = "automl")
}
\arguments{
\item{path}{A string specifying the directory path where the model is saved. Default is './'.}

\item{filename}{A string specifying the name of the saved model file. Default is 'automl'.}
}
\value{
The loaded H2O model object.
}
\description{
\code{nm_load_h2o} loads a previously saved H2O model from disk.
}
\details{
This function loads the H2O model from the specified directory and filename.
}
\examples{
\dontrun{
h2o.init()
model <- nm_load_h2o(path = './models', filename = 'my_model')
}
}
