% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_pdp.r
\name{nm_pdp}
\alias{nm_pdp}
\title{Generate Partial Dependence Plots (PDP)}
\usage{
nm_pdp(
  df,
  model,
  variables = NULL,
  training_only = TRUE,
  grid.resolution = 20,
  n_cores = NULL
)
}
\arguments{
\item{df}{Data frame containing the input data.}

\item{model}{The trained model object.}

\item{variables}{A vector of feature names for which PDPs are to be generated. Default is NULL (all feature names will be used).}

\item{training_only}{Logical indicating whether to use only training data for generating PDPs. Default is TRUE.}

\item{grid.resolution}{The number of points to evaluate on the grid for each feature. Default is 20.}

\item{n_cores}{Number of CPU cores to use for parallel processing. Default is system's total minus one.}
}
\value{
A data frame containing the partial dependence values.
}
\description{
\code{nm_pdp} generates partial dependence plots for specified features using a trained model.
}
\examples{
\dontrun{
library(dplyr)
df <- data.frame(
  set = rep(c("train", "test"), each = 50),
  feature1 = rnorm(100),
  feature2 = rnorm(100)
)
model <- lm(feature1 ~ feature2, data = df)
pdp_results <- nm_pdp(df, model, varibales = c("feature1", "feature2"))
}
}
