% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_pdp.r
\name{nm_pdp_worker}
\alias{nm_pdp_worker}
\title{Worker function for generating PDP}
\usage{
nm_pdp_worker(model, df, variable, grid.resolution)
}
\arguments{
\item{model}{The trained model object.}

\item{df}{Data frame containing the input data.}

\item{variable}{The feature name for which PDP is to be generated.}

\item{grid.resolution}{The number of points to evaluate on the grid for the feature.}
}
\value{
A data frame containing the partial dependence values for the specified feature.
}
\description{
\code{nm_pdp_worker} is a worker function that generates partial dependence values for a specific feature.
}
\examples{
\dontrun{
library(dplyr)
library(pdp)
df <- data.frame(
  set = rep(c("train", "test"), each = 50),
  feature1 = rnorm(100),
  feature2 = rnorm(100)
)
model <- lm(feature1 ~ feature2, data = df)
pdp_result <- nm_pdp_worker(df, model, feature = "feature2")
}
}
