% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_prepare_train_model.r
\name{nm_prepare_train_model}
\alias{nm_prepare_train_model}
\title{Prepare and Train Model}
\usage{
nm_prepare_train_model(
  df,
  value,
  feature_names,
  split_method,
  fraction,
  model_config,
  seed,
  verbose = TRUE
)
}
\arguments{
\item{df}{Data frame containing the input data.}

\item{value}{The target variable name as a string.}

\item{feature_names}{The names of the features used for training.}

\item{split_method}{The method for splitting data into training and testing sets.}

\item{fraction}{The proportion of the data to be used for training.}

\item{model_config}{A list containing configuration parameters for model training.}

\item{seed}{A random seed for reproducibility.}

\item{verbose}{Should the function print progress messages? Default is TRUE.}
}
\value{
The trained leader model from AutoML.
}
\description{
\code{nm_prepare_train_model} prepares the data and trains a model using AutoML.
}
\examples{
\dontrun{
library(dplyr)
df <- data.frame(
  feature1 = rnorm(100),
  feature2 = rnorm(100),
  target = rnorm(100)
)
res <- nm_prepare_train_model(df, value = "target", feature_names = c("feature1", "feature2"), model_type = "lm")
}
}
