% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_prepare_data.r
\name{nm_process_df}
\alias{nm_process_df}
\title{Process DataFrame for Model Training}
\usage{
nm_process_df(df, variables_col)
}
\arguments{
\item{df}{Input data frame.}

\item{variables_col}{A vector of column names to be selected along with the date column.}
}
\value{
Processed data frame with selected columns.
}
\description{
\code{nm_process_df} processes the input DataFrame by checking for date in row names, identifying
the date column, and selecting relevant features.
}
\examples{
\dontrun{
df <- data.frame(
  feature1 = rnorm(100),
  feature2 = rnorm(100),
  date = Sys.time() + seq(1, 100, by = 1)
)
processed_df <- nm_process_df(df, variables_col = c("feature1", "feature2"))
}
}
