% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_rolling.r
\name{nm_rolling}
\alias{nm_rolling}
\title{Apply rolling window meteorological normalisation}
\usage{
nm_rolling(
  df = NULL,
  model = NULL,
  value = "value",
  feature_names = NULL,
  variables_resample = NULL,
  split_method = "random",
  fraction = 0.75,
  model_config = NULL,
  n_samples = 300,
  window_days = 14,
  rolling_every = 7,
  seed = 7654321,
  n_cores = NULL,
  memory_save = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{df}{Data frame containing the input data.}

\item{model}{Pre-trained model for normalisation. If not provided, a model will be trained.}

\item{value}{The target variable name as a string.}

\item{feature_names}{The names of the features used for normalisation.}

\item{split_method}{The method for splitting data into training and testing sets. Default is 'random'.}

\item{fraction}{The proportion of the data to be used for training. Default is 0.75.}

\item{model_config}{A list containing configuration parameters for model training.}

\item{n_samples}{Number of times to sample the data for normalisation. Default is 300.}

\item{window_days}{The size of the rolling window in days. Default is 14.}

\item{rolling_every}{The interval at which the rolling window is applied in days. Default is 7.}

\item{seed}{A random seed for reproducibility. Default is 7654321.}

\item{n_cores}{Number of CPU cores to use for parallel processing. Default is system's total minus one.}

\item{memory_save}{Logical indicating whether to save memory by processing each sample independently.
If \code{TRUE}, resampling and prediction are done in memory-efficient batches. If \code{FALSE}, all samples
are generated and processed at once, which uses more memory. Default is FALSE.}

\item{verbose}{Should the function print progress messages? Default is TRUE.}
}
\value{
A list containing the normalised data frame and model statistics.
}
\description{
\code{nm_rolling} performs meteorological normalisation on data using a rolling window approach.
}
\examples{
\dontrun{
library(dplyr)
library(h2o)
library(purrr)
df <- data.frame(
  date = Sys.time() + seq(1, 100, by = 1),
  feature1 = rnorm(100),
  feature2 = rnorm(100),
  target = rnorm(100)
)
result <- nm_rolling(df, value = "target", feature_names = c("feature1", "feature2"), variables_resample = c("feature1"), window_size = 10, model_type = "h2o.gbm", seed = 12345)
}
}
