% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_save_h2o.r
\name{nm_save_h2o}
\alias{nm_save_h2o}
\title{Save Trained H2O Model}
\usage{
nm_save_h2o(model, path = "./", filename = "automl")
}
\arguments{
\item{model}{The trained H2O model object.}

\item{path}{A string specifying the directory path where the model will be saved. Default is './'.}

\item{filename}{A string specifying the desired filename for the saved model. Default is 'automl'.}
}
\value{
A string indicating the path of the saved model.
}
\description{
\code{nm_save_h2o} saves a trained H2O model to a specified directory.
}
\details{
This function ensures the output directory exists, saves the H2O model, and renames
the saved model file to the desired filename.
}
\examples{
\dontrun{
h2o.init()
model <- h2o.gbm(x = c('feature1', 'feature2'), y = 'outcome', training_frame = df)
nm_save_h2o(model, path = './models', filename = 'my_model')
}
}
