% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_scm.r
\name{nm_scm_parallel}
\alias{nm_scm_parallel}
\title{Main function to perform SCM in parallel}
\usage{
nm_scm_parallel(
  df,
  poll_col,
  date_col,
  code_col,
  control_pool,
  post_col,
  training_split = 0.75,
  n_cores = NULL
)
}
\arguments{
\item{df}{Data frame containing the input data.}

\item{poll_col}{The name of the column containing the pollutant data.}

\item{date_col}{The name of the column containing the date data.}

\item{code_col}{The name of the column containing the unit codes.}

\item{control_pool}{A vector of codes representing the control units.}

\item{post_col}{The name of the column indicating the post-treatment period.}

\item{training_split}{The proportion of the data to be used for training. Default is 0.75.}

\item{n_cores}{Number of CPU cores to use for parallel processing. Default is system's total minus one.}
}
\value{
A data frame with the actual and synthetic control data for all treatment targets.
}
\description{
\code{nm_scm_parallel} applies the synthetic control method in parallel for multiple treatment targets.
}
\examples{
\dontrun{
library(dplyr)
library(parallel)
df <- data.frame(
  date = Sys.time() + seq(1, 100, by = 1),
  pollutant = rnorm(100),
  unit_code = rep(c("A", "B", "C", "D"), each = 25),
  post = rep(c(0, 1), each = 50)
)
result <- nm_scm_parallel(df, poll_col = "pollutant", date_col = "date",
                          code_col = "unit_code", treat_targets = c("A"),
                          control_pool = c("B", "C", "D"), post_col = "post",
                          n_cores = 2)
}
}
