% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_modStats.r
\name{nm_Stats}
\alias{nm_Stats}
\title{Calculate Various Statistics}
\usage{
nm_Stats(df, mod, obs, statistic = NULL)
}
\arguments{
\item{df}{Data frame containing the input data.}

\item{mod}{The name of the model predictions column.}

\item{obs}{The name of the observed values column.}

\item{statistic}{A vector of statistics to calculate. Default is c("n", "FAC2", "MB", "MGE", "NMB", "NMGE", "RMSE", "r", "COE", "IOA", "R2").}
}
\value{
A data frame containing the calculated statistics.
}
\description{
\code{nm_Stats} calculates various statistical measures for model evaluation.
}
\examples{
\donttest{
library(dplyr)
# Example usage
df <- data.frame(value_predict = rnorm(100), value = rnorm(100))
stats <- nm_Stats(df, mod = "value_predict", obs = "value")
}

}
