% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_train_model.r
\name{nm_train_model}
\alias{nm_train_model}
\title{Train a model using H2O's AutoML}
\usage{
nm_train_model(
  df,
  value = "value",
  variables = NULL,
  model_config = NULL,
  seed = 7654321,
  n_cores = NULL,
  verbose = TRUE,
  max_retries = 3
)
}
\arguments{
\item{df}{Input data frame containing the data to be used for training.}

\item{value}{The target variable name as a string. Default is "value".}

\item{variables}{Independent/explanatory variables used for training the model.}

\item{model_config}{A list containing configuration parameters for model training. If not provided, defaults will be used.}

\item{seed}{A random seed for reproducibility. Default is 7654321.}

\item{n_cores}{Number of CPU cores to use for the model training. Default is system's total minus one.}

\item{verbose}{Should the function print progress messages? Default is TRUE.}

\item{max_retries}{Number of times to retry training the model if a connection error occurs. Default is 3.}
}
\value{
The trained AutoML model.
}
\description{
\code{nm_train_model} is a function to train a model using H2O's AutoML.
It initializes H2O, checks for duplicate and missing variables, extracts relevant data for training,
sets up parallel processing, and trains the model using AutoML.
}
\examples{
\donttest{
# Load necessary libraries
library(h2o)
library(dplyr)

# Prepare example data
data_example <- data.frame(
  value = rnorm(100),
  var1 = rnorm(100),
  var2 = rnorm(100),
  set = rep(c("training", "testing"), 50)
)

# Train AutoML model using the example data
model <- nm_train_model(
  df = data_example,
  value = "value",
  variables = c("var1", "var2"),
  model_config = list(max_models = 5, time_budget = 600)
)
}
}
